Release v2.7.2 (Compilation error fix)

+++ lib/sbi/nnrf-handler.c
@@ -866,9 +866,9 @@ static void handle_validity_time(
     }
 
     ogs_info("[%s] Subscription %s until %s "
-            "[duration:%ld,validity:%d.%06d,patch:%d.%06d]",
+            "[duration:%lld,validity:%d.%06d,patch:%d.%06d]",
             subscription_data->id, action, validity_time_string,
-            subscription_data->validity_duration,
+            (long long)subscription_data->validity_duration,
             (int)ogs_time_sec(subscription_data->validity_duration),
             (int)ogs_time_usec(subscription_data->validity_duration),
             (int)ogs_time_sec(patch), (int)ogs_time_usec(patch));
